+++ src/translation/english.c
@@ -402,7 +402,7 @@ static translation_string all_strings[] = {
     {TR_TOOLTIP_BUTTON_STOCKPILING_OFF, "Disable stockpiling"},
     {TR_BUILDING_LIGHTHOUSE, "Lighthouse"},
     {TR_BUILDING_CARAVANSERAI, "Caravanserai"},
-    {TR_BUILDING_CARAVANSERAI_DESC, "Merchant caravans shelter here to rest and feed their animals. Land trade disruptions last half as long."},
+    {TR_BUILDING_CARAVANSERAI_DESC, "Caravanserai is fully operational. Merchant caravans shelter here to rest and feed their animals. Land trade disruptions last half as long."},
     {TR_CARAVANSERAI_SPECIAL_ORDERS_HEADER, "Caravanserai instructions"},
     {TR_BUILDING_CARAVANSERAI_PHASE_1, "(Foundation)"},
     {TR_BUILDING_CARAVANSERAI_PHASE_2, "(Colonnade)"},
@@ -421,6 +421,7 @@ static translation_string all_strings[] = {
     {TR_BUILDING_CARAVANSERAI_POLICY_3, "Landed quantity policy: Trade caravans carry 4 additional goods, but the Caravanserai consumes 20% more food." },
     {TR_BUILDING_CARAVANSERAI_NO_EMPLOYEES, "The Caravanserai has no employees. Food can't be collected." },
     {TR_BUILDING_CARAVANSERAI_NO_FOOD, "The Caravanserai has employees, but they are currently looking for an available source of food." },
+    {TR_BUILDING_CARAVANSERAI_FOOD_SHORTAGE, "The Caravanserai has employees and food, but the stocks are too low to supply all merchant caravans." },
     {TR_BUILDING_LIGHTHOUSE_PHASE_1, "(Foundation)" },
     {TR_BUILDING_LIGHTHOUSE_PHASE_2, "(Pedestal)" },
     {TR_BUILDING_LIGHTHOUSE_PHASE_3, "(Tower)" },
@@ -429,7 +430,8 @@ static translation_string all_strings[] = {
     {TR_BUILDING_LIGHTHOUSE_PHASE_2_TEXT, "Masons are constructing the pedestal that will raise the Lighthouse's flame far above the horizon." },
     {TR_BUILDING_LIGHTHOUSE_PHASE_3_TEXT, "The Lighthouse rises higher each day, as masons lay successive courses of the great tower." },
     {TR_BUILDING_LIGHTHOUSE_PHASE_4_TEXT, "Artisans are putting the final touches on the construction of the Lighthouse. Soon it will shine for miles, guiding ships home." },
-    {TR_BUILDING_LIGHTHOUSE_BONUS_DESC, "The light atop this great tower guides ships from the far reaches of the world safely to your city. Fishing ships now move 10% faster, and sea storms last half as long." },
+    {TR_BUILDING_LIGHTHOUSE_BONUS_DESC, "Lighthouse is fully operational. The light atop this great tower guides ships from the far reaches of the world safely to your city. Fishing ships now move 10% faster, and sea storms last half as long." },
+    {TR_BUILDING_LIGHTHOUSE_NO_TIMBER, "Lightkeepers are struggling to find the timber necessary to keep the flame alight. The ships approaching your shores strain to see the fading beacon."},
     {TR_BUILDING_LIGHTHOUSE_POLICY_TITLE, "Sea trade policy" },
     {TR_BUILDING_LIGHTHOUSE_NO_POLICY, "No sea trade policy." },
     {TR_BUILDING_LIGHTHOUSE_POLICY_TEXT, "Select a sea trade policy for your city. You can change to another policy at any time." },
