+++ src/translation/translation.h
@@ -417,6 +417,7 @@ typedef enum {
     TR_BUILDING_CARAVANSERAI_POLICY_3,
     TR_BUILDING_CARAVANSERAI_NO_EMPLOYEES,
     TR_BUILDING_CARAVANSERAI_NO_FOOD,
+    TR_BUILDING_CARAVANSERAI_FOOD_SHORTAGE,
     TR_BUILDING_LIGHTHOUSE_PHASE_1,
     TR_BUILDING_LIGHTHOUSE_PHASE_2,
     TR_BUILDING_LIGHTHOUSE_PHASE_3,
@@ -426,6 +427,7 @@ typedef enum {
     TR_BUILDING_LIGHTHOUSE_PHASE_3_TEXT,
     TR_BUILDING_LIGHTHOUSE_PHASE_4_TEXT,
     TR_BUILDING_LIGHTHOUSE_BONUS_DESC,
+    TR_BUILDING_LIGHTHOUSE_NO_TIMBER,
     TR_BUILDING_LIGHTHOUSE_POLICY_TITLE,
     TR_BUILDING_LIGHTHOUSE_NO_POLICY,
     TR_BUILDING_LIGHTHOUSE_POLICY_TEXT,
